Update volume slider using previous value & increment.
This is what should be the "correct implementation" of that feature, however in the past Chrome would behave very weirdly and necessitate setting it to min/max to work at all. This behavior seems to have been fixed in recent Chrome versions (or Spotify changed something), so we can now use code that makes intuitive sense.

Fixes #19. Alternative to #20 (thanks Joshua!).

Co-authored-by: Joshua Pritt <39056491+prittjr@users.noreply.github.com>